Move password string validation into PasswordInput
This was StringUtils's last function so it was removed
