--- ui/src/Utils/StringUtils.ts
@@ -1,28 +1,0 @@
-/*
- * Copyright (c) 2022 Ford Motor Company
- * All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-export const UPPERCASE_REGEX = /[A-Z]/;
-export const NUMBER_REGEX = /\d/;
-
-export function validatePasswordString(password: string): boolean {
-	return (
-		!!password &&
-		password.length >= 8 &&
-		!!password.match(UPPERCASE_REGEX) &&
-		!!password.match(NUMBER_REGEX)
-	);
-}
